Extended ParamBulder to be included in an entire API or namespace.
