+++ lib/grape.rb
@@ -26,7 +26,7 @@ require 'virtus'
 I18n.load_path << File.expand_path('../grape/locale/en.yml', __FILE__)
 
 module Grape
-  extend ActiveSupport::Autoload
+  extend ::ActiveSupport::Autoload
 
   eager_autoload do
     autoload :API
@@ -47,14 +47,14 @@ module Grape
   end
 
   module Http
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     eager_autoload do
       autoload :Headers
     end
   end
 
   module Exceptions
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :Base
     autoload :Validation
     autoload :ValidationArrayErrors
@@ -78,22 +78,27 @@ module Grape
   end
 
   module Extensions
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
 
     autoload :DeepMergeableHash
     autoload :DeepSymbolizeHash
     autoload :Hash
-    autoload :HashWithIndifferentAccess
+
+    module ActiveSupport
+      extend ::ActiveSupport::Autoload
+
+      autoload :HashWithIndifferentAccess
+    end
 
     module Hashie
-      extend ActiveSupport::Autoload
+      extend ::ActiveSupport::Autoload
 
       autoload :Mash
     end
   end
 
   module Middleware
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :Base
     autoload :Versioner
     autoload :Formatter
@@ -102,7 +107,7 @@ module Grape
     autoload :Stack
 
     module Auth
-      extend ActiveSupport::Autoload
+      extend ::ActiveSupport::Autoload
       autoload :Base
       autoload :DSL
       autoload :StrategyInfo
@@ -110,7 +115,7 @@ module Grape
     end
 
     module Versioner
-      extend ActiveSupport::Autoload
+      extend ::ActiveSupport::Autoload
       autoload :Path
       autoload :Header
       autoload :Param
@@ -119,7 +124,7 @@ module Grape
   end
 
   module Util
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :InheritableValues
     autoload :StackableValues
     autoload :ReverseStackableValues
@@ -129,7 +134,7 @@ module Grape
   end
 
   module ErrorFormatter
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :Base
     autoload :Json
     autoload :Txt
@@ -137,7 +142,7 @@ module Grape
   end
 
   module Formatter
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :Json
     autoload :SerializableHash
     autoload :Txt
@@ -145,13 +150,13 @@ module Grape
   end
 
   module Parser
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :Json
     autoload :Xml
   end
 
   module DSL
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     eager_autoload do
       autoload :API
       autoload :Callbacks
@@ -170,17 +175,17 @@ module Grape
   end
 
   class API
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :Helpers
   end
 
   module Presenters
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :Presenter
   end
 
   module ServeFile
-    extend ActiveSupport::Autoload
+    extend ::ActiveSupport::Autoload
     autoload :FileResponse
     autoload :FileBody
     autoload :SendfileResponse
