+++ lib/grape/extensions/active_support/hash_with_indifferent_access.rb
@@ -1,0 +1,25 @@
+module Grape
+  module Extensions
+    module ActiveSupport
+      module HashWithIndifferentAccess
+        extend ::ActiveSupport::Concern
+
+        included do
+          namespace_inheritable(:build_params_with, Grape::Extensions::ActiveSupport::HashWithIndifferentAccess::ParamBuilder)
+        end
+
+        def params_builder
+          Grape::Extensions::ActiveSupport::HashWithIndifferentAccess::ParamBuilder
+        end
+
+        module ParamBuilder
+          def build_params
+            params = ::ActiveSupport::HashWithIndifferentAccess[rack_params]
+            params.deep_merge!(grape_routing_args) if env[Grape::Env::GRAPE_ROUTING_ARGS]
+            params
+          end
+        end
+      end
+    end
+  end
+end
