+++ lib/grape/request.rb
@@ -5,7 +5,7 @@ module Grape
     alias rack_params params
 
     def initialize(env, options = {})
-      extend options[:build_params_with] || Grape::Extensions::HashWithIndifferentAccess::ParamBuilder
+      extend options[:build_params_with] || Grape::Extensions::ActiveSupport::HashWithIndifferentAccess::ParamBuilder
       super(env)
     end
